+++ CHANGELOG.md
@@ -10,15 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added in version 1.0.0
 
 * Initial release
-* Wrote documentation
 * Implemented the honeypot using the Twisted framework
-* Make the honeypot compatible with Python 3.x
+* Made the honeypot compatible with Python 3.x
 * Config file support
 * Various command-line options
 * Log rotation
 * Support for the `report_public_ip` config file option
 * A script for starting, stopping, and restarting the honeypot
 * Macros like `$ip`, `$now`, `$old`
+* Documentation
 * Emulation of the following IPP operations:
   * `Get-Printer-Attributes`
   * `Get-Jobs`
@@ -26,8 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * `Print-Job`
 * Output plugin support
 * Output plugins for
+  * CouchDB
   * Elasticsearch
   * HPFeeds
+  * Influx 2.0 (Python 3 only)
   * JSON
   * MongoDB
   * MySQL